prepare and reorganize tests around remote urls
