+++ test/remote_url/html_import_remote_css/main.html
@@ -1,0 +1,23 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <title>Title</title>
+    <meta charset="utf-8" />
+    <!-- <link rel="icon" href="data:," /> -->
+    <!-- TODO later: add an integrity attribute -->
+    <link
+      rel="stylesheet"
+      type="text/css"
+      href="http://localhost:9999/roboto.css"
+    />
+    <style>
+      body {
+        font-family: "Roboto";
+      }
+    </style>
+  </head>
+
+  <body>
+    Hello world
+  </body>
+</html>
