+++ test/remote_url/js_import_remote_js/main.js
