+++ test/remote_url/js_import_remote_js/server/file.js
